pwmの誤解を解き大幅に変更

+++ train/src/main/java/com/traindiorama/pulse/MortorData.java
@@ -1,14 +1,14 @@
 package com.traindiorama.pulse;
 
 public interface MortorData {
-    public long MaxPulseRangeMicroseconds = 2500;
-    public long MinPulseRangeMicroseconds = 500;
+    public int MaxPulseRangeMicroseconds = 2500;
+    public int MinPulseRangeMicroseconds = 500;
 
-    public long NeutralPulseMicroseconds = 1500;
+    public int NeutralPulseMicroseconds = 1500;
 
-    public long NeutralPeriodMicroseconds = 2000;
+    public int NeutralPeriodMicroseconds = 2000;
 
-    public short Angle_Degree_MaxRangePulse = 270;
-    public short Angle_Degree_NeutralPulse = 135;
-    public short Angle_Degree_MinRangePulse = 0;
+    public int Angle_Degree_MaxRangePulse = 270;
+    public int Angle_Degree_NeutralPulse = 135;
+    public int Angle_Degree_MinRangePulse = 0;
 }
